% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise_matrix.R
\name{add_noise_matrix}
\alias{add_noise_matrix}
\title{add random or targeted noise to a non-invertible matrix}
\usage{
add_noise_matrix(
  A = NULL,
  noise.threshold = c(0, 1),
  noise.type = c("random", "targeted")
)
}
\arguments{
\item{A}{numeric matrix}

\item{noise.threshold}{vector of two values, minimum and maximum value for the perturbation}

\item{noise.type}{random (to any column) or targeted (to columns that are linear combinations of others)}
}
\value{
perturbed matrix
}
\description{
add random or targeted noise to a non-invertible matrix
}
