% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_block_matrix.R
\name{build_block_matrix}
\alias{build_block_matrix}
\title{build an empty block matrix}
\usage{
build_block_matrix(plants, herbivores, floral.visitors)
}
\arguments{
\item{plants}{character vector}

\item{herbivores}{character vector}

\item{floral.visitors}{character vector}
}
\value{
block matrix with appropriate names, NA in their elements
}
\description{
build the skeleton of a block matrix given
names of plants, herbivores, and floral visitors
}
