% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_block_matrix.R
\name{fill_block_matrix}
\alias{fill_block_matrix}
\title{Fill a block matrix}
\usage{
fill_block_matrix(
  block.matrix,
  pp.matrix = NA,
  ph.matrix = NA,
  pfv.matrix = NA,
  fv.overlap.matrix = NA,
  h.overlap.matrix = NA,
  switch.herb.sign = TRUE
)
}
\arguments{
\item{block.matrix}{empty block matrix}

\item{pp.matrix}{plant-plant neighbours}

\item{ph.matrix}{plant-herbivore matrix}

\item{pfv.matrix}{plant-floral visitor matrix}

\item{fv.overlap.matrix}{floral visitor intraguild matrix}

\item{h.overlap.matrix}{herbivore intraguild matrix}

\item{switch.herb.sign}{if TRUE, plants have different signs over herbivores than herbivores over plants}
}
\value{
a filled block matrix of the same dimensions as block.matrix
}
\description{
fill the elements of an empty block matrix (build_block_matrix) given
a series of intra- and interguild matrices
}
