% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrix_feasibility.R
\name{get_matrix_feasibility}
\alias{get_matrix_feasibility}
\title{Get the feasibility domain of a given matrix}
\usage{
get_matrix_feasibility(
  A = NULL,
  bootstrap.replicates = 1000,
  noise.threshold = c(0, 1),
  noise.type = c("random", "targeted")
)
}
\arguments{
\item{A}{interaction matrix}

\item{bootstrap.replicates}{replicates in case A is not invertible}

\item{noise.threshold}{numeric vector with a minimum and a maximum amount of noise
to be incorporated in case A is not invertible}

\item{noise.type}{'random' or 'targeted'. Noise added to any column at random, or
to columns which are linear combinations of the rest}
}
\value{
dataframe with feasibility domain, the condition number (see base function 'kappa'),
and other info if A is not invertible.
}
\description{
this is encapsulated to account for cases in which the matrix is not invertible,
and we add a small amount of noise
}
