% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phenologic_overlap.R
\name{get_phenologic_overlap}
\alias{get_phenologic_overlap}
\title{Calculate phenologic overlap}
\usage{
get_phenologic_overlap(sp.data)
}
\arguments{
\item{sp.data}{dataframe with the following columns: ID, year,
min.month, min.day, max.month, max.day.
This dataframe can be built from the function 'filter_sp_taxonomy_phenology'}
}
\value{
numeric matrix
}
\description{
returns a numeric matrix, with each element being the overlap
between sp i and sp j relative to the complete phenological period of sp i
}
