% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preference_matrix.R
\name{get_preference_matrix}
\alias{get_preference_matrix}
\title{Get species preferences for a series of matrices}
\usage{
get_preference_matrix(matrix.list)
}
\arguments{
\item{matrix.list}{list of matrices, potentially nested}
}
\value{
list with the same structure as matrix.list
}
\description{
Species preferences as defined in Staniczenko et al. (2013) Nat. Comm
This function is similar in structure to 'get_overlap_matrix'
}
