% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomic_overlap.R
\name{get_taxonomic_overlap}
\alias{get_taxonomic_overlap}
\title{calculate binary taxonomic overlap
returns a binary matrix with 1 if two species are from the same order
and 0 otherwise}
\usage{
get_taxonomic_overlap(sp.data)
}
\arguments{
\item{sp.data}{dataframe with the following columns: sp.ID, order, year,
min.month, min.day, max.month, max.day.
This dataframe can be built from the function 'filter_sp_taxonomy_phenology'}
}
\value{
binary (0-1) matrix
}
\description{
calculate binary taxonomic overlap
returns a binary matrix with 1 if two species are from the same order
and 0 otherwise
}
