% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_sp.R
\name{get_valid_sp}
\alias{get_valid_sp}
\title{get the valid names from a series of guild matrices}
\usage{
get_valid_sp(pp.matrix = NULL, ph.matrix = NULL, pfv.matrix = NULL)
}
\arguments{
\item{pp.matrix}{plant-plant matrix}

\item{ph.matrix}{plant-herbivores matrix}

\item{pfv.matrix}{plant-floral visitors matrix}
}
\value{
list with three character vectors, one per guild
}
\description{
extracts the names of sp that will form the row
built by combining these three matrices. This is because the different bipartite
matrices may not be 100\% concordant in their species. So, this serves as a kind
of filter
}
