% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horizontal_community_matrix.R
\name{horizontal_community_matrix}
\alias{horizontal_community_matrix}
\title{Create a matrix for a horizontal community network}
\usage{
horizontal_community_matrix(
  S = 5,
  c = 0.5,
  tau = 1.5,
  min.diag.dom = 0,
  restricted.positive = TRUE,
  int.mean = 0,
  int.sd = 1
)
}
\arguments{
\item{S}{richness}

\item{c}{connectance}

\item{tau}{inverse of kurtosis (tau ~1.5 gives kurtosis ~3 for S=25)
see https://statisticaloddsandends.wordpress.com/2019/04/15/the-sinh-arcsinh-normal-distribution/}

\item{min.diag.dom}{the minimum diagonal dominance wanted in each row}

\item{restricted.positive}{whether interactions are only >0 or not}

\item{int.mean}{mean of interaction strength (taken from a normal distribution)}

\item{int.sd}{standard deviation of interaction strength}
}
\value{
square matrix
}
\description{
Create a matrix for a horizontal community network
}
