% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraguild_out_strength.R
\name{intraguild_out_strength}
\alias{intraguild_out_strength}
\title{species-level metrics}
\usage{
intraguild_out_strength(A, sp.names)
}
\arguments{
\item{A}{interaction matrix}

\item{sp.names}{which sp to calculate, must be in names of A}
}
\value{
dataframe
}
\description{
different functions with the same structure, metric is in function name
}
