% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr_modularity.R
\name{lr_modularity}
\alias{lr_modularity}
\title{LinkRank Modularity}
\usage{
lr_modularity(g, partition, damping = 0.85, pr.algo = "prpack", weights = NULL)
}
\arguments{
\item{g}{graph (igraph object)}

\item{partition}{graph partition (numeric vector of memberships or "communities" object)}

\item{damping}{damping factor (1 - teleportation prob.)}

\item{pr.algo}{algorithm to calculate Perron vector,
possible options are "prpack", "arpack", and "power"}

\item{weights}{If this is NULL and the graph has a weight edge attribute
then that is used. If weights is a numerical vector then
it used, even if the graph has a weights edge attribute.
If this is NA, then no edge weights are used (even if the
graph has a weight edge attribute)}
}
\value{
numeric value
}
\description{
The modified modularity is related to the
random walk process in the network, and the global meaning
of the modified modularity is the fraction of time spent moving
within communities by a random walker minus the expected
value of this fraction. And locally, the meaning of the
modified modularity is that a link with higher LinkRank is
more likely to be assigned as an intracommunity link than a
link with lower LinkRank. The definition of community is
also changed, according to the change in modularity. In this
definition, a community is a group of nodes in which a random
walker is more likely to stay.
}
\details{
Reproduced from https://github.com/baruuum/LinkRank_Modularity
Source:
Kim, Y., S.W. Son, H. Jeong. 2010. "Finding Communities in Directed Networks",
Physical Review E, 81
}
