% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/May_complexity.R
\name{May_complexity}
\alias{May_complexity}
\title{May's complexity metric}
\usage{
May_complexity(A = NULL, S = NULL, C = NULL, sigma = NULL)
}
\arguments{
\item{A}{numeric matrix. Either A or S,C,sigma must be provided}

\item{S}{optional, species richness}

\item{C}{optional, connectance}

\item{sigma}{optional, standard deviation of interaction strengths}
}
\value{
numeric value
}
\description{
Calculates the classic complexity metric from May (1972) for a given -unipartite- interaction matrix
or connectance, richness, and interaction strength values.
}
