% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_matrix.R
\name{normalize_matrix}
\alias{normalize_matrix}
\title{Normalize a series of matrices in a list}
\usage{
normalize_matrix(matrix.list = NULL, max.value = NULL)
}
\arguments{
\item{matrix.list}{list of matrices, potentially nested}

\item{max.value}{optional value for normalizing}
}
\value{
list with the same structure as matrix.list, containing normalized matrices
}
\description{
Normalize according to a provided max.value or according
to the maximum value observed across all matrices in the list
}
