% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega.R
\name{Omega}
\alias{Omega}
\title{normalized feasibility}
\usage{
Omega(alpha)
}
\arguments{
\item{alpha}{interaction matrix}
}
\value{
numeric value
}
\description{
these functions are taken from
"A guideline to study the feasibility domain of multi-trophic and changing ecological communities" by:
Chuliang Song, Rudolf P. Rohr, Serguei Saavedra
published in: Journal of Theoretical Biology
}
