% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Omega_bootstrap.R
\name{Omega_bootstrap}
\alias{Omega_bootstrap}
\title{Bootstrap version of feasibility calculation}
\usage{
Omega_bootstrap(alpha, replicates = 100)
}
\arguments{
\item{alpha}{interaction matrix}

\item{replicates}{number of bootstrap replicates}
}
\value{
normalized feasibility
}
\description{
Motivated because the original function 'Omega' is
not deterministic
}
