% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_variability.R
\name{relative_variability}
\alias{relative_variability}
\title{Title
Uses Heath (2006) method to compute temporal variability in population abundances,
and compares it to the coefficient of variation.
DOI:10.1111/j.2006.0030-1299.15067.x}
\usage{
relative_variability(x, Diff = NULL)
}
\arguments{
\item{x}{vector of population densities}

\item{Diff}{}
}
\value{
list with two values
}
\description{
Title
Uses Heath (2006) method to compute temporal variability in population abundances,
and compares it to the coefficient of variation.
DOI:10.1111/j.2006.0030-1299.15067.x
}
