% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_metrics.R
\name{species_metrics}
\alias{species_metrics}
\title{wrapper function for several species-level metrics}
\usage{
species_metrics(
  A,
  metrics = c("intraguild_out_strength", "interguild_out_strength",
    "intraguild_in_strength", "interguild_in_strength", "intraguild_intraspecific"),
  sp.names
)
}
\arguments{
\item{A}{interaction matrix}

\item{metrics}{character vector, including one or more of the following:
"intraguild_out_strength",
"interguild_out_strength",
"intraguild_in_strength",
"interguild_in_strength",
"intraguild_intraspecific"}

\item{sp.names}{which species to calculate, must be named in A}
}
\value{
dataframe
}
\description{
calculates a series of metrics from a set of species in an interaction matrix
}
