% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere_sampling.R
\name{sphere_sampling}
\alias{sphere_sampling}
\title{sample random growth rates}
\usage{
sphere_sampling(m, positive.r = NULL)
}
\arguments{
\item{m}{number of sp}

\item{positive.r}{positions of sp with strictly positive growth rate}
}
\value{
vector of sampled growth rates
}
\description{
these functions are taken from
"A guideline to study the feasibility domain of multi-trophic and changing ecological communities" by:
Chuliang Song, Rudolf P. Rohr, Serguei Saavedra
published in: Journal of Theoretical Biology
}
