% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectance.R
\name{connectance}
\alias{connectance}
\title{connectance function for unipartite or bipartite matrices}
\usage{
connectance(interaction.matrix, quant = FALSE)
}
\arguments{
\item{interaction.matrix}{interaction matrix}

\item{quant}{whether to calculate quantitative connectance after LDq' in Banasek-Richter et al. 2009}
}
\value{
connectance value (0-1)
}
\description{
either qualitative or quantitative
}
